Rename `buildJs` into `assembleJs`
