+++ buildSrc/src/main/kotlin/io/spine/internal/gradle/javascript/task/impl/Report.kt
@@ -29,16 +29,17 @@ package io.spine.internal.gradle.javascript.task.impl
 import io.spine.internal.gradle.javascript.task.JsTaskRegistering
 import io.spine.internal.gradle.report.license.generateLicenseReport
 
+// TODO:2021-09-22:alexander.yevsyukov: Resolve this dependency.
+// generateLicenseReport.finalizedBy npmLicenseReport
+
+/**
+ * Registers a [task][io.spine.internal.gradle.javascript.task.npmLicenseReport] for creating
+ * a license report for the module's `NPM` dependencies.
+ */
 fun JsTaskRegistering.report() {
-
-    // TODO:2021-09-22:alexander.yevsyukov: Resolve this dependency.
-    // generateLicenseReport.finalizedBy npmLicenseReport
-
-    generateLicenseReport.configure {
-        finalizedBy(
-            npmLicenseReport()
-        )
-    }
+    generateLicenseReport.finalizedBy(
+        npmLicenseReport()
+    )
 }
 
 private fun JsTaskRegistering.npmLicenseReport() =
@@ -48,6 +49,10 @@ private fun JsTaskRegistering.npmLicenseReport() =
         group = jsAnyTask
 
         doLast {
+
+            // The script below generates license report for NPM dependencies and appends it
+            // to the report for Java dependencies generated by `generateLicenseReport` task.
+
             npm("run", "license-report")
         }
     }
